Add files via upload
These are the changes that I have been working on. Some of them are still in progress, but it would be best to push it now in case anybody wants to help review it. Added ceremony and name endpoints. Also updated app.js and mongo.js to run it properly. Sorry I couldn't finish some of the methods; I got my second COVID shot this morning and it hit me hard in the evening.