Add files via upload
Two endpoints that focus on names and ceremony number. Some of the methods are still in progress, but it would be best to upload it now in case anybody wants to review it. Sorry I couldn't finish them all; I got the second COVID vaccination shot this morning and it hit me in the evening badly.